- Work around three bugs in capidl affecting Sleep.idl.
- Sleep_sleepTill was confusing nanoseconds with milliseconds. Fixed.

- Combine act_WakeUpIn with act_WakeUpAtTick.

- act_Enqueue now relies on caller to do any locking necessary for the queue.

- Fix a bug in printf("%lld").

+++ src/sys/kernel/kern_printf.c
@@ -360,7 +360,7 @@ snumbers:
         ull = va_arg(ap, unsigned long long);
         if ((long)ull < 0) {
           neg = 1;
-          ull = -(long)ull;
+          ull = -(long long)ull;
         }
         goto printull;
     }
